+++ src/main/java/blackjack/domain/participant/Participant.java
@@ -3,6 +3,11 @@ package blackjack.domain.participant;
 import blackjack.domain.Money;
 import blackjack.domain.carddeck.Card;
 import blackjack.domain.state.*;
+import blackjack.domain.state.finished.Blackjack;
+import blackjack.domain.state.finished.Bust;
+import blackjack.domain.state.finished.Stay;
+import blackjack.domain.state.hand.Hand;
+import blackjack.domain.state.hand.Score;
 
 import java.util.Collections;
 import java.util.List;
