+++ src/main/java/blackjack/domain/state/State.java
@@ -1,6 +1,8 @@
 package blackjack.domain.state;
 
 import blackjack.domain.carddeck.Card;
+import blackjack.domain.state.hand.Hand;
+import blackjack.domain.state.hand.Score;
 
 import java.util.List;
