+++ src/main/java/blackjack/domain/state/finished/Finished.java
@@ -1,6 +1,9 @@
-package blackjack.domain.state;
+package blackjack.domain.state.finished;
 
 import blackjack.domain.carddeck.Card;
+import blackjack.domain.state.hand.Hand;
+import blackjack.domain.state.Started;
+import blackjack.domain.state.State;
 
 public abstract class Finished extends Started {
