+++ src/main/java/blackjack/domain/state/running/Running.java
@@ -1,4 +1,7 @@
-package blackjack.domain.state;
+package blackjack.domain.state.running;
+
+import blackjack.domain.state.hand.Hand;
+import blackjack.domain.state.Started;
 
 public abstract class Running extends Started {
