+++ src/test/java/blackjack/domain/state/HandTest.java
@@ -3,6 +3,8 @@ package blackjack.domain.state;
 import blackjack.domain.carddeck.Card;
 import blackjack.domain.carddeck.Number;
 import blackjack.domain.carddeck.Pattern;
+import blackjack.domain.state.hand.Hand;
+import blackjack.domain.state.hand.Score;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
