+++ src/test/java/blackjack/domain/state/HitTest.java
@@ -3,6 +3,10 @@ package blackjack.domain.state;
 import blackjack.domain.carddeck.Card;
 import blackjack.domain.carddeck.Number;
 import blackjack.domain.carddeck.Pattern;
+import blackjack.domain.state.finished.Bust;
+import blackjack.domain.state.finished.Stay;
+import blackjack.domain.state.hand.Hand;
+import blackjack.domain.state.running.Hit;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
