+++ src/test/java/blackjack/domain/state/StateFactoryTest.java
@@ -3,6 +3,8 @@ package blackjack.domain.state;
 import blackjack.domain.carddeck.Card;
 import blackjack.domain.carddeck.Number;
 import blackjack.domain.carddeck.Pattern;
+import blackjack.domain.state.finished.Blackjack;
+import blackjack.domain.state.running.Hit;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
